Change Cronet network thread priority to int
For those wondering "who in their right mind could possibly want to use
floating point for thread priority", the answer is "iOS". But Cronet
doesn't support iOS anymore, so let's change this to int to reduce
eyebrow strain.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5496764
Auto-Submit: Etienne Dechamps <edechamps@google.com>
Commit-Queue: Etienne Dechamps <edechamps@google.com>
Cr-Commit-Position: refs/heads/main@{#1295480}

+++ components/cronet/url_request_context_config.h
@@ -167,7 +167,7 @@ struct URLRequestContextConfig {
 
   // Optional network thread priority.
   // On Android, corresponds to android.os.Process.setThreadPriority() values.
-  const std::optional<double> network_thread_priority;
+  const std::optional<int> network_thread_priority;
 
   // Whether the connection status of active bidirectional streams should be
   // monitored.
@@ -211,7 +211,7 @@ struct URLRequestContextConfig {
       // Optional network thread priority.
       // On Android, corresponds to android.os.Process.setThreadPriority()
       // values. Do not specify for other targets.
-      std::optional<double> network_thread_priority);
+      std::optional<int> network_thread_priority);
 
  private:
   URLRequestContextConfig(
@@ -245,7 +245,7 @@ struct URLRequestContextConfig {
       // Optional network thread priority.
       // On Android, corresponds to android.os.Process.setThreadPriority()
       // values. Do not specify for other targets.
-      std::optional<double> network_thread_priority);
+      std::optional<int> network_thread_priority);
 
   // Parses experimental options from their JSON format to the format used
   // internally.
@@ -322,7 +322,7 @@ struct URLRequestContextConfigBuilder {
   // Optional network thread priority.
   // On Android, corresponds to android.os.Process.setThreadPriority() values.
   // Do not specify for other targets.
-  std::optional<double> network_thread_priority;
+  std::optional<int> network_thread_priority;
 };
 
 }  // namespace cronet
